unknown google sheets network on index.html. advanced.html does not work.

+++ index.js
@@ -1,20 +1,4 @@
-require("main.js");
-
-function require(script) {
-    $.ajax({
-        url: script,
-        dataType: "script",
-        async: false,           // <-- This is the key
-        success: function () {
-            // all good...
-        },
-        error: function () {
-            throw new Error("Could not load script " + script);
-        }
-    });
-}
-
-
+import * as main from "/main.js";
 
 const searchBox = document.getElementById("search-box");
 searchBox.addEventListener("keyup", function (event) {
@@ -24,24 +8,30 @@ searchBox.addEventListener("keyup", function (event) {
 });
 
 function simpleSearchFunction() {
-
+    if(main.isSearchError())return;
     if (isSearchBroken) {
-      searchIsBroken();
+      main.searchIsBroken();
       return;
     }
     else if (!isSearchReady) {
-      searchNotReady();
+      main.searchNotReady();
       return;
     }
   
     main.deleteAllTiles();
   
-    var searchBoxValue = document.getElementById("search-box").value;
+    var searchBoxValue = searchBox.value;
     var searchForPhrases = [];
     searchForPhrases[1]=searchBoxValue;//extra weight is given to exact term instead of individual parts
   
     //complecated regex to split string into string array based on spaces outside of quotes.
-    searchBoxValue.match(/(".*?"|[^" \s]+)(?=\s* |\s*$)/g).forEach(function (val, index) { searchForPhrases[index] = val.replaceAll("\"", "") });
+    searchBoxValue.match(/(".*?"|[^" \s]+)(?=\s* |\s*$)/g).
+        forEach(
+            function (val, index) { 
+                searchForPhrases[index] = val.replaceAll("\"", "") 
+            }
+        );
+    //purposfuly starts at 1, not zero. first element is already set.
     for (var i = 1; i < searchForPhrases.length; i++) {
       searchForPhrases[i] = searchForPhrases[i].trim();
     }
@@ -55,14 +45,12 @@ function simpleSearchFunction() {
     var numberOfTiles = 0;
     for (let i = 0; i < serviceData.length; i++) {
     if (serviceData[i].searchRating>0) {
-      renderOneTile(serviceData[i].title, serviceData[i].description, serviceData[i].minAge, serviceData[i].address, serviceData[i].website, serviceData[i].zipcode, serviceData[i].tags, false);
+      main.renderOneTile(serviceData[i].title, serviceData[i].description, serviceData[i].minAge, serviceData[i].address, serviceData[i].website, serviceData[i].zipcode, serviceData[i].tags, false);
       numberOfTiles++;
     }
     }
     if (numberOfTiles == 0) {
-      let clone = nothingFoundTemplate.content.cloneNode(true);
-      var myNode = document.getElementById("groupDiv");
-      myNode.appendChild(clone);
+        main.renderNothingFoundCard();
     }
     document.documentElement.scrollTop = document.getElementById("groupDiv").getBoundingClientRect().top-4;
